requirements.txt
streamlit
hopsworks
pandas
scikit-learn
shap
matplotlib
seaborn
python-dotenv
numpy
joblib